Ensure the systemd file is created under the correct name
When installing mongodb-org packages on Debian/Ubuntu the daemon name is
set to "mongod" but the .service file is unconditionally installed as
"mongodb.service". This PR ensures, that the systemd file (and the
accompanying symlink) are installed as "mongodb_dameon_name.service".
Otherwise installation fails with
fatal: [tmongo-02]: FAILED! => {"changed": false, "msg": "Could not find the requested service mongod: host"}
for every host, as the handler tries to restart mongod, but systemd only
sees mongodb